requirements.txt
Django
psycopg2
uWSGI
Pillow
xlwt
celery
redis
django-filter
django-admin-autocomplete-all